Database code updated

+++ app/src/main/java/uk/ac/tees/live/q5273477/journal/Entry.java
@@ -15,6 +15,6 @@ public class Entry {
     public int _id;
     public String text;
     public String categoty;
-    public Float date_time;
+    public Long date_time;
 
 }
